zadacha5

+++ 2025/24311002/05_Money_in_your_pocket/05_Money_in_your_pocket.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int m05,m10,m20,m50,m1,m2;
+    float m05,m10,m20,m50,m1,m2;
     cout << "Enter coins: 05, 10, 20, 50, 100 and 200" << endl;
     cin >> m05 >> m10 >> m20 >> m50 >> m1 >> m2;
